tests: Extract out test_message_dict.py for MessageDict tests.
